+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added a `keep_truth` argument to the supervised simulation functions. This retains the column that defines the error free simulated value of the outcome. This numeric column is called `.truth`. 
 
+* A new simulation function, `sim_logistic()` was added. 
+
 # modeldata 1.0.1
 
 * Small update to fix HTML  for CRAN.